+++ __tests__/fixtures/plugin-foo.mjs
@@ -1,3 +1,5 @@
+import path from 'node:path';
+
 import stylelint from 'stylelint';
 
 const {
@@ -9,20 +11,46 @@ const ruleName = 'plugin/foo';
 
 const messages = ruleMessages(ruleName, {
 	rejected: (selector) => `No "${selector}" selector`,
+	expectFilename: (expected, actual) => `Expect "${actual}" to be "${expected}"`,
 });
 
 /** @type {(value: unknown) => boolean} */
 const isString = (value) => typeof value === 'string';
 
 /** @type {import('stylelint').Rule} */
-const ruleFunction = (primary) => {
+const ruleFunction = (primary, secondaryOptions) => {
 	return (root, result) => {
-		const validOptions = validateOptions(result, ruleName, {
-			actual: primary,
-			possible: [isString],
-		});
+		const validOptions = validateOptions(
+			result,
+			ruleName,
+			{
+				actual: primary,
+				possible: [isString],
+			},
+			{
+				actual: secondaryOptions,
+				possible: {
+					filename: [isString],
+				},
+				optional: true,
+			},
+		);
 
 		if (!validOptions) {
+			return;
+		}
+
+		const expectedFilename = secondaryOptions?.filename;
+		const actualFilename = path.basename(root.source?.input.file ?? '');
+
+		if (expectedFilename && expectedFilename !== actualFilename) {
+			report({
+				result,
+				ruleName,
+				message: messages.expectFilename(expectedFilename, actualFilename),
+				node: root,
+			});
+
 			return;
 		}