+++ __tests__/getTestRule.test.mjs
@@ -47,6 +47,20 @@ testRule({
 	accept: [{ code: '.a {}' }],
 });
 
+testRule({
+	plugins,
+	ruleName,
+	config: ['.a', { filename: 'foo.css' }],
+	accept: [{ code: '.a {}', codeFilename: 'foo.css' }],
+	reject: [
+		{
+			code: '.a {}',
+			codeFilename: 'bar.css',
+			message: messages.expectFilename('foo.css', 'bar.css'),
+		},
+	],
+});
+
 const testRuleWithLoadLint = getTestRule({
 	loadLint: () => import('stylelint').then((m) => m.default.lint),
 });