+++ index.d.ts
@@ -5,6 +5,11 @@ export type TestCase = {
 	code: string;
 
 	/**
+	 * The filename for this `code` property.
+	 */
+	codeFilename?: string;
+
+	/**
 	 * Description of the test case.
 	 */
 	description?: string;