+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Head
+
+- Added: `codeFilename` option to `accept`/`rejected` cases in `testRule` function.
+
 ## 7.0.1
 
 - Fixed: missing `funding` field in `package.json`.